Organized everything into a script that can be ran inside of update_table_scripts.  Later down the road we can put this on a schedule but for now we can run manually to update when we want.
